Extract src/models/provider_profiles/ cluster (#191)
Move ProviderProfile and its three credential sub-record models
(ProviderLicensure, ProviderEducation, ProviderCertification) plus the
two colocated test files under `src/models/provider_profiles/`. The four
models already form a cluster: the profile owns the three sub-records
via `cascade="all, delete-orphan"` + FK CASCADE, and all four share the
same table-CHECK pattern against tuples in `src/models/enums.py`.

This is the second of two extractions for #191. After this PR the model
layer's domain clusters (`posts/`, `provider_profiles/`) are explicit;
parent-level files are limited to genuinely shared utilities (`base.py`,
`enums.py`) and singletons (`user.py`, `audit_log.py`).

External `from src.models import …` is unchanged: the package `__init__`
re-exports from the new locations.

Also declare `selectolax` as an `additional_dependencies` entry on the
title-case-check pre-commit hook. #198 made the import a hard failure
when missing; without this declaration, pre-commit's isolated venv
(which does not see project deps from `pyproject.toml`) blocks every
commit that touches a `.md` / `.html` / `.jinja*` file.
