+++ src/models/provider_profiles/provider_profile.py
@@ -2,8 +2,8 @@ from sqlalchemy import CheckConstraint, Column, ForeignKey, Text
 from sqlalchemy.orm import relationship
 from sqlalchemy.types import Uuid
 
-from .base import BaseModel
-from .enums import LOCATION_AVAILABILITY_OPTIONS, US_STATES, check_in_tuple_sql
+from ..base import BaseModel
+from ..enums import LOCATION_AVAILABILITY_OPTIONS, US_STATES, check_in_tuple_sql
 
 _TABLE = "provider_profiles"
 
@@ -20,7 +20,7 @@ class ProviderProfile(BaseModel):
     """Provider's directory profile — practice info, location, and session
     availability flags. A user may own zero, one, or many profiles; ownership
     is recorded via the non-unique `user_id` FK. Distinct from
-    `provider_availability_details` (which is a per-post detail row for a
+    `posts/provider_availability_detail.py` (which is a per-post detail row for a
     specific outreach `Post`); a `ProviderProfile` is a long-lived directory
     entry that owns the provider's credential lists (licensures, educations,
     certifications) via cascade.
